Separate file for graph list and handle case of no selected graph
