+++ frog/imports/ui/svg/index.js
@@ -2,15 +2,12 @@ import React, { Component } from 'react';
 import { Provider } from 'mobx-react';
 import Form from 'react-jsonschema-form';
 import styled from 'styled-components';
-import { createContainer } from 'meteor/react-meteor-data';
-
-import { removeGraph } from '../../api/activities';
-import { Graphs, addGraph } from '../../api/graphs';
 
 import { connect, store } from './store';
 import Graph from './Graph';
 import Rename from './Rename';
 import SidePanel from './SidePanel';
+import { assignGraph, GraphList } from './GraphList';
 
 import './App.css';
 
@@ -42,36 +39,6 @@ const SettingsContainer = styled.div`
   position: absolute;
   top: 900px;
 `;
-
-const GL = createContainer(
-  props => ({ ...props, graphs: Graphs.find().fetch() }),
-  ({ graphs, graphId }) => (
-    <div>
-      <h3>Graph list</h3>
-      <button
-        className="btn btn-primary btn-sm"
-        onClick={() => store.setId(addGraph())}
-      >
-        New
-      </button>
-      <ul>
-        {graphs.length ? graphs.map(graph => (
-              <li style={{ listStyle: 'none' }} key={graph._id}>
-                <a href="#" onClick={() => removeGraph(graph._id)}>
-                  <i className="fa fa-times" />
-                </a>
-                <a href="#" onClick={() => store.setId(graph._id)}>
-                  <i className="fa fa-pencil" />
-                </a>
-                {graph._id} {graph._id === graphId ? ' (current)' : null}
-              </li>
-            )) : <li>No graph</li>}
-      </ul>
-    </div>
-  )
-);
-
-const GraphList = connect(({ store: { id } }) => <GL graphId={id} />);
 
 const App = connect(({ store: { panOffset, hasSelection } }) => (
   <div>
@@ -109,7 +76,7 @@ const App = connect(({ store: { panOffset, hasSelection } }) => (
 
 export default class AppClass extends Component {
   componentWillMount() {
-    store.setId(1);
+    assignGraph();
   }
 
   render() {
